+++ srcs/parser/parser.h
@@ -6,7 +6,7 @@
 /*   By: tasano <tasano@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/09 23:54:22 by tasano            #+#    #+#             */
-/*   Updated: 2022/12/27 22:24:47 by tasano           ###   ########.fr       */
+/*   Updated: 2022/12/28 11:08:53 by tasano           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,26 +17,36 @@
 #include <unistd.h>
 #include "lexer.h"
 
-typedef struct s_redirection
+typedef struct s_redirect
 {
 	char                    *filename;
     t_token_type             type;
     t_token_type            expnad_type;
-    struct s_redirection    *next;
-} t_redirection;
+    struct s_redirect    *next;
+} t_redirect;
 
 typedef struct s_cmd
 {
-    t_redirection   *output;
-    struct s_cmd    *piped_cmd;
-    t_redirection   *input;
-    char            **cmd;
-    size_t          argc;
+    t_redirect   *output;
+    struct s_cmd *piped_cmd;
+    t_redirect   *input;
+    char         **cmd;
+    size_t       argc;
 }   t_cmd;
 
 t_cmd   *parser(t_token_lst *lst);
-t_cmd	*cmd_addback(t_cmd *lst, t_cmd *new);
-void	redirection_addback(t_redirection **lst, t_redirection *new);
-t_redirection	*redirection_new(t_token_lst *lst);
-t_cmd	*cmd_new();
+
+//redirect lst util
+void	    redirection_addback(t_redirect **lst, t_redirect *new);
+t_redirect	*redirection_new(t_token_lst *lst);
+void        redirect_lstfree(t_redirect **lst);
+
+//cmd lst utils
+void	    cmd_lstfree(t_cmd **cmd);
+t_cmd	    *cmd_addback(t_cmd *lst, t_cmd *new);
+t_cmd	    *cmd_new(); 
+//utils
+char        **append_args(char **args, size_t argc, char *new);
+int         parse_error(t_token_lst *lst, t_cmd *cmd, t_token_lst **tmplst);
+void        free_parser_lst(t_token_lst **lst);
 #endif
